Check system dependencies necessary for package 'glatos'

Checks to make sure external dependencies required by functions in 'glatos' are installed and available to R.

check_dependencies()

Value

results of checks to your terminal

Details

check_dependencies checks that the Geospatial Data Abstraction Library (GDAL) and ffmpeg (cross platform software for manipulating video content) software are installed on your computer and accessible to R. GDAL is used by the make_transition function to create a transition layer required by interpolate_path for non-linear interpolation. FFmpeg is required to create or modify video animations of fish movement using the make_frames, make_video, and adjust_video_playback functions.

When check_dependencies is executed, R attempts to sequentially access the external libraries. If the libraries are installed and accessible, a message is returned to the terminal stating that the check was successful. Failed attempts to access the external libraries are printed to the terminal.

Installation of the GDAL library and a number of other open-source programs useful for working with spatial data on windows is accomplished by downloading the network installer for your appropriate windows computer (32 or 64 bit) at https://trac.osgeo.org/osgeo4w/. Installation of the GDAL library on Mac is possible using Homebrew or KyngChaos 3rd party repositories. Alternatively, the GDAL library is incorporated in QGIS (open source desktop GIS) software and may be obtained by installing QGIS. Standalone installers for QGIS are available (Windows, Mac, and Linux) at https://qgis.org/en/site/forusers/download.html.

Installation of the ffmpeg library on windows is accomplished by downloading the recent 'static' build from http://ffmpeg.zeranoe.com/builds/. After the download is complete, use your favorite compression utility to extract the downloaded folder. Decompress the package and store contents on your computer. Last, Edit your system path variable to include the path to the directory containing ffmpeg.exe

Installation of ffmpeg on Mac is similar to windows. First, download most recent build from http://www.evermeet.cx/ffmpeg/. The binary files are compressed with 7zip so may need toinstall an unarchiving utility (http://wakaba.c3.cx/s/apps/unarchiver.html) to extract the program folder. After the folder is extracted, copy the ffmpeg folder to /usr/local/bin/ffmpeg on your machine.

Examples

# run check
check_dependencies()

#> Checking for gdal...

#> OK... gdal version 2.2.1 is installed

#> Checking for ffmpeg...

#> OK... FFmpeg is installed and on system PATH.