An R package for the Great Lakes Acoustic Telemetry Observation System

glatos is an R package with functions useful to members of the Great Lakes Acoustic Telemetry Observation System (http://glatos.glos.us). Functions may be generally useful for processing, analyzing, simulating, and visualizing acoustic telemetry data, but are not strictly limited to acoustic telemetry applications.

Package status

This package is in early development and its content is evolving. To access the package or contribute code, join the project at (https://gitlab.oceantrack.org/GreatLakes/glatos). If you encounter problems or have questions or suggestions, please post a new issue or email <cholbrook@usgs.gov> (maintainer: Chris Holbrook).

Installation

Installation instructions can be found at https://gitlab.oceantrack.org/GreatLakes/glatos/wikis/installation-instructions.

Data loading and processing

read_glatos_detections and read_otn_detections

Fast data loading from standard GLATOS and OTN data files to a single structure that is compatible with other glatos functions.

read_glatos_receivers

Reads receiver location histories from standard GLATOS data files to a single structure that is compatible with other glatos functions.

read_glatos_workbook

Reads project-specific receiver history and fish taggging and release data from a standard glatos workbook file.

read_vemco_tag_specs

Reads transmitter (tag) specifications and operating schedule.

real_sensor_values

Converts 'raw' transmitter sensor (e.g., depth, temperature) to 'real'-scale values (e.g., depth in meters) using transmitter specification data (e.g., from read\_vemco\_tag\_specs).

Filtering and summarizing

min_lag

Facilitates identification and removal of false positive detections by calculating the minimum time interval (min\_lag) between successive detections.

false_detections

Removes potential false positive detections using "short interval" criteria (see min_lag).

detection_events

Distills detection data down to a much smaller number of discrete detection events, defined as a change in location or time gap that exceeds a threshold.

summarize_detections

Calculates number of fish detected, number of detections, first and last detection timestamps, and/or mean location of receivers or groups, depending on specific type of summary requested.

Visualization and data exploration

abacus_plot

Useful for exploring movement patterns of individual tagged animals through time.

detection_bubble_plot

Useful for exploring distribution of tagged individuals among receivers.

interpolate_path, make_frames, and make_video

Interpolate spatio-temporal movements, between detections, create video frames, and stitch frames together to create animated video file using FFmpeg software.

adjust_playback_time

Modify playback speed of videos and optionally convert between video file formats. Requires FFmpeg.

Simulation functions for system design and evaluation

calc_collision_prob

Estimates the probability of collisions for pulse-position-modulation type co-located telemetry transmitters. This is useful for determining the number of fish to release or tag specifications (e.g., delay).

receiver_line_det_sim

Simulates detection of acoustic-tagged fish crossing a receiver line (or single receiver). This is useful for determining optimal spacing of receviers in a line and tag specifications (e.g., delay).

crw_in_polygon, transmit_along_path, and detect_transmissions

Individually simulate random fish movement paths within a water body (crw_in_polygon: a random walk in a polygon), tag signal transmissions along those paths (transmit_along_path: time series and locations of transmissions based on tag specs), and detection of those transmittions by receivers in a user-defined receiver network (detect_transmissions: time series and locations of detections based on detection range curve). Collectively, these functions can be used to explore, compare, and contrast theoretical performance of a wide range of transmitter and receiver network designs.