Simulate detection of transmitter signals in a receiver network

Simulates detection of transmitter signals in a receiver network based on detection range curve (detection probability as a function of distance), location of transmitter, and location of receivers.

detect_transmissions(trnsLoc = NA, recLoc = NA, detRngFun = NA)

Arguments

trnsLoc
A three-column data frame with locations (numeric columns named 'x' and 'y') and timestamps (numeric or POSIXct column named 'et') where signals were transmitted.

recLoc
A two-column data frame with receiver locations (numeric columns named 'x' and 'y')

detRngFun
A function that defines detection range curve; must accept a numeric vector of distances and return a numeric vector of detection probabilities at each distance.

Value

A data frame containing:

trns_id

Unique signal transmission ID

recv_id

Unique receiver ID

recv_x

Receiver x coordinate

recv_y

Receiver y coordinate

trns_x

Transmitter x coordinate

trns_y

Transmitter y coordinate

etime

Elapsed time

Details

Distances between each signal transmission location and receiver are calculated using pythagorean theorem. The probability of detecting each signal on each receiver is determined from the detection range curve. Detection of each signal on each receiver is determined stochastically by draws from a Bernoulli distribution with probability p (detection prob).

This function was written to be used along with transmit_along_path.

See also

Examples

#make a simple path in polygon
mypath <- crw_in_polygon(data.frame(x = c(0, 0, 1000, 1000),
  y = c(0, 1000, 1000, 0)), stepLen=100, nsteps=50)

#>   |                                                                              |                                                                      |   0%

#> Error in (heading_deg - 180):(heading_deg + 180): argument of length 0

plot(mypath,type='l',xlim=c(0,1000),ylim=c(0,1000)) #view path

#> Error in plot(mypath, type = "l", xlim = c(0, 1000), ylim = c(0, 1000)): object 'mypath' not found


#add receivers
recs <- expand.grid(c(250,750),c(250,750))
names(recs) <- c("x","y") #needed by detect_transmissions
points(recs, pch=15, col="blue")

#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet


#simulate tag transmissions
mytrns <- transmit_along_path(mypath,vel=2.0,delayRng=c(60,180),burstDur=5.0)

#> Error in diff(path$x): object 'mypath' not found

points(mytrns,pch=21) #add to plot

#> Error in points(mytrns, pch = 21): object 'mytrns' not found


#Define detection range function (to pass as detRngFun) 
# that returns detection probability for given distance
# assume logistic form of detection range curve where 
#   dm = distance in meters
#   b = intercept and slope
pdrf <- function(dm, b=c(0.5, -1/120)){
  p <- 1/(1+exp(-(b[1]+b[2]*dm)))
  return(p)
}
pdrf(c(100,200,300,400,500)) #view detection probs. at some distances

#> [1] 0.41742979 0.23745803 0.11920292 0.05554926 0.02492443


#simulate detection
mydtc <- detect_transmissions(trnsLoc=mytrns, recLoc=recs, detRngFun=pdrf)

#> Error in as.vector(y): object 'mytrns' not found

#view transmissions that were detected
points(trns_y~trns_x, data=mydtc,pch=21, bg="red")

#> Error in eval(m$data, eframe): object 'mydtc' not found