Generate the residence index from a set of detections

This residence index tool will take a compressed or uncompressed detection file and caculate the residency index for each station/receiver in the detections. The information passed to the function is what is used to calculate the residence index, make sure you are only passing the data you want taken into consideration for the residence index (i.e. species, stations, tags, etc.).

residence_index(detections, calculation_method = "kessel")

Arguments

detections
A data.frame from the detection_events function.

calculation_method
A character string with the calculation method using one of the following: kessel, timedelta, aggregate_with_overlap, or aggregate_no_overlap

Value

A data.frame of days_detected, residency_index, location, mean_latitude, mean_longitude

Details

The Kessel method converts both the first_detection and last_detection columns into a date with no hours, minutes, or seconds. Next it creates a list of the unique days where a detection was seen. The size of the list is returned as the total number of days as an integer. This calculation is used to determine the total number of distinct days (T) and the total number of distinct days per station (S). Possible rounding error may occur as a detection on 2016-01-01 23:59:59 and a detection on 2016-01-02 00:00:01 would be counted as two days when it is really 2-3 seconds.

$$ RI = S/T$$ $$ RI = Residence Index$$ $$S = Distinct number of days detected at the station$$ $$T = Distinct number of days detected anywhere on the array$$

The Timedelta calculation method determines the first detection and the last detection of all detections. The time difference is then taken as the values to be used in calculating the residence index. The timedelta for each station is divided by the timedelta of the array to determine the residence index.

$$ RI = Delta S/Delta T$$

$$RI = Residence Index$$

$$Delta S = Last detection time at a station - First detection time at the station$$

$$Delta T = Last detection time on an array - First detection time on the array$$

The Aggregate With Overlap calculation method takes the length of time of each detection and sums them together. A total is returned. The sum for each station is then divided by the sum of the array to determine the residence index.

$$ RI = AwOS/AwOT$$

$$RI = Residence Index$$

$$AwOS = Sum of length of time of each detection at the station$$

$$AwOT = Sum of length of time of each detection on the array$$

The Aggregate No Overlap calculation method takes the length of time of each detection and sums them together. However, any overlap in time between one or more detections is excluded from the sum. For example, if the first detection is from 2016-01-01 01:02:43 to 2016-01-01 01:10:12 and the second detection is from 2016-01-01 01:09:01 to 2016-01-01 01:12:43, then the sum of those two detections would be 10 minutes. A total is returned once all detections of been added without overlap. The sum for each station is then divided by the sum of the array to determine the residence index.

$$ RI = AnOS/AnOT$$

$$RI = Residence Index$$

$$AnOS = Sum of length of time of each detection at the station, excluding any overlap$$

$$AnOT = Sum of length of time of each detection on the array, excluding any overlap$$

References

Kessel, S.T., Hussey, N.E., Crawford, R.E., Yurkowski, D.J., O'Neill, C.V. and Fisk, A.T., 2016. Distinct patterns of Arctic cod (Boreogadus saida) presence and absence in a shallow high Arctic embayment, revealed across open-water and ice-covered periods through acoustic telemetry. Polar Biology, 39(6), pp.1057-1068. https://www.researchgate.net/publication/279269147

Examples

#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
det <- read_glatos_detections(det_file)
detection_events <- glatos::detection_events(det)

#> The event filter distilled 7180 detections down to 165 distinct detection events.

rik_data <- glatos::residence_index(detection_events, calculation_method = 'kessel')
rit_data <- glatos::residence_index(detection_events, calculation_method = 'timedelta')
riawo_data <- glatos::residence_index(detection_events, calculation_method = 'aggregate_with_overlap')
riano_data <- glatos::residence_index(detection_events, calculation_method = 'aggregate_no_overlap')